added 2 more mockito tests for Post Request

+++ src/main/java/com/tsi/sjumbe/demo/Film.java
@@ -40,7 +40,7 @@ public class Film implements Serializable {
                 String description,
                 String rating,
                 int release_year,
-                int length,
+                float length,
                 int language_id){
 
         this.title = title;
