+++ src/main/java/com/tsi/sjumbe/demo/SakilaMicroserviceApplication.java
@@ -88,7 +88,7 @@ public class SakilaMicroserviceApplication {
 
 
 	@PostMapping("/Add_Film")
-	public @ResponseBody String addFilm(@RequestParam String title, String description, String rating, int release_year, int length, int language_id ){
+	public @ResponseBody String addFilm(@RequestParam String title, String description, String rating, int release_year, float length, int language_id ){
 		Film addFilm = new Film(title,description,rating,release_year,length,language_id);
 		filmRepository.save(addFilm);
 		return save;
